Reorganize code to extract shared supabase realtime subscription logic
