+++ app/features/receive/cashu-token-swap-hooks.ts
@@ -1,8 +1,5 @@
 import type { Token } from '@cashu/cashu-ts';
-import {
-  REALTIME_SUBSCRIBE_STATES,
-  type RealtimePostgresChangesPayload,
-} from '@supabase/supabase-js';
+import type { RealtimePostgresChangesPayload } from '@supabase/supabase-js';
 import {
   type QueryClient,
   useMutation,
@@ -10,7 +7,8 @@ import {
   useQuery,
   useQueryClient,
 } from '@tanstack/react-query';
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo } from 'react';
+import { useSupabaseRealtimeSubscription } from '~/lib/supabase/supabase-realtime';
 import { useLatest } from '~/lib/use-latest';
 import { useGetLatestCashuAccount } from '../accounts/account-hooks';
 import {
@@ -192,15 +190,6 @@ function useCompleteCashuTokenSwap() {
   });
 }
 
-type SubscriptionState =
-  | {
-      status: 'subscribing' | 'subscribed' | 'closed';
-    }
-  | {
-      status: 'error';
-      error: Error;
-    };
-
 function useOnCashuTokenSwapChange({
   onCreated,
   onUpdated,
@@ -208,67 +197,37 @@ function useOnCashuTokenSwapChange({
   onCreated: (swap: CashuTokenSwap) => void;
   onUpdated: (swap: CashuTokenSwap) => void;
 }) {
-  const [state, setState] = useState<SubscriptionState>({
-    status: 'subscribing',
-  });
   const cashuCryptography = useCashuCryptography();
   const onCreatedRef = useLatest(onCreated);
   const onUpdatedRef = useLatest(onUpdated);
 
-  useEffect(() => {
-    const channel = agicashDb
-      .channel('cashu-token-swaps')
-      .on(
-        'postgres_changes',
-        {
-          event: '*',
-          schema: 'wallet',
-          table: 'cashu_token_swaps',
-        },
-        async (
-          payload: RealtimePostgresChangesPayload<AgicashDbCashuTokenSwap>,
-        ) => {
-          if (payload.eventType === 'INSERT') {
-            const swap = await CashuTokenSwapRepository.toTokenSwap(
-              payload.new,
-              cashuCryptography.decrypt,
-            );
-            onCreatedRef.current(swap);
-          } else if (payload.eventType === 'UPDATE') {
-            const swap = await CashuTokenSwapRepository.toTokenSwap(
-              payload.new,
-              cashuCryptography.decrypt,
-            );
-            onUpdatedRef.current(swap);
-          }
-        },
-      )
-      .subscribe((status, error) => {
-        if (status === REALTIME_SUBSCRIBE_STATES.SUBSCRIBED) {
-          setState({ status: 'subscribed' });
-        } else if (status === REALTIME_SUBSCRIBE_STATES.CLOSED) {
-          setState({ status: 'closed' });
-        } else {
-          setState({
-            status: 'error',
-            error: new Error(
-              `Error with "${channel.topic}" channel subscription. Status: ${status}`,
-              { cause: error },
-            ),
-          });
+  return useSupabaseRealtimeSubscription(() =>
+    agicashDb.channel('cashu-token-swaps').on(
+      'postgres_changes',
+      {
+        event: '*',
+        schema: 'wallet',
+        table: 'cashu_token_swaps',
+      },
+      async (
+        payload: RealtimePostgresChangesPayload<AgicashDbCashuTokenSwap>,
+      ) => {
+        if (payload.eventType === 'INSERT') {
+          const swap = await CashuTokenSwapRepository.toTokenSwap(
+            payload.new,
+            cashuCryptography.decrypt,
+          );
+          onCreatedRef.current(swap);
+        } else if (payload.eventType === 'UPDATE') {
+          const swap = await CashuTokenSwapRepository.toTokenSwap(
+            payload.new,
+            cashuCryptography.decrypt,
+          );
+          onUpdatedRef.current(swap);
         }
-      });
-
-    return () => {
-      channel.unsubscribe();
-    };
-  }, [cashuCryptography]);
-
-  if (state.status === 'error') {
-    throw state.error;
-  }
-
-  return state.status;
+      },
+    ),
+  );
 }
 
 function usePendingCashuTokenSwaps() {
