+++ app/features/send/cashu-send-quote-hooks.ts
@@ -1,8 +1,5 @@
 import { type MeltQuoteResponse, MintOperationError } from '@cashu/cashu-ts';
-import {
-  REALTIME_SUBSCRIBE_STATES,
-  type RealtimePostgresChangesPayload,
-} from '@supabase/supabase-js';
+import type { RealtimePostgresChangesPayload } from '@supabase/supabase-js';
 import {
   type QueryClient,
   useMutation,
@@ -13,6 +10,7 @@ import type Big from 'big.js';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { getCashuUnit, getCashuWallet } from '~/lib/cashu';
 import type { Money } from '~/lib/money';
+import { useSupabaseRealtimeSubscription } from '~/lib/supabase/supabase-realtime';
 import {
   type LongTimeout,
   clearLongTimeout,
@@ -235,15 +233,6 @@ export function useCashuSendQuote({
   };
 }
 
-type SubscriptionState =
-  | {
-      status: 'subscribing' | 'subscribed' | 'closed';
-    }
-  | {
-      status: 'error';
-      error: Error;
-    };
-
 function useOnCashuSendQuoteChange({
   onCreated,
   onUpdated,
@@ -251,67 +240,37 @@ function useOnCashuSendQuoteChange({
   onCreated: (send: CashuSendQuote) => void;
   onUpdated: (send: CashuSendQuote) => void;
 }) {
-  const [state, setState] = useState<SubscriptionState>({
-    status: 'subscribing',
-  });
   const cashuCryptography = useCashuCryptography();
   const onCreatedRef = useLatest(onCreated);
   const onUpdatedRef = useLatest(onUpdated);
 
-  useEffect(() => {
-    const channel = agicashDb
-      .channel('cashu-send-quotes')
-      .on(
-        'postgres_changes',
-        {
-          event: '*',
-          schema: 'wallet',
-          table: 'cashu_send_quotes',
-        },
-        async (
-          payload: RealtimePostgresChangesPayload<AgicashDbCashuSendQuote>,
-        ) => {
-          if (payload.eventType === 'INSERT') {
-            const addedQuote = await CashuSendQuoteRepository.toSend(
-              payload.new,
-              cashuCryptography.decrypt,
-            );
-            onCreatedRef.current(addedQuote);
-          } else if (payload.eventType === 'UPDATE') {
-            const updatedQuote = await CashuSendQuoteRepository.toSend(
-              payload.new,
-              cashuCryptography.decrypt,
-            );
-            onUpdatedRef.current(updatedQuote);
-          }
-        },
-      )
-      .subscribe((status, error) => {
-        if (status === REALTIME_SUBSCRIBE_STATES.SUBSCRIBED) {
-          setState({ status: 'subscribed' });
-        } else if (status === REALTIME_SUBSCRIBE_STATES.CLOSED) {
-          setState({ status: 'closed' });
-        } else {
-          setState({
-            status: 'error',
-            error: new Error(
-              `Error with "${channel.topic}" channel subscription. Status: ${status}`,
-              { cause: error },
-            ),
-          });
+  return useSupabaseRealtimeSubscription(() =>
+    agicashDb.channel('cashu-send-quotes').on(
+      'postgres_changes',
+      {
+        event: '*',
+        schema: 'wallet',
+        table: 'cashu_send_quotes',
+      },
+      async (
+        payload: RealtimePostgresChangesPayload<AgicashDbCashuSendQuote>,
+      ) => {
+        if (payload.eventType === 'INSERT') {
+          const addedQuote = await CashuSendQuoteRepository.toSend(
+            payload.new,
+            cashuCryptography.decrypt,
+          );
+          onCreatedRef.current(addedQuote);
+        } else if (payload.eventType === 'UPDATE') {
+          const updatedQuote = await CashuSendQuoteRepository.toSend(
+            payload.new,
+            cashuCryptography.decrypt,
+          );
+          onUpdatedRef.current(updatedQuote);
         }
-      });
-
-    return () => {
-      channel.unsubscribe();
-    };
-  }, [cashuCryptography]);
-
-  if (state.status === 'error') {
-    throw state.error;
-  }
-
-  return state.status;
+      },
+    ),
+  );
 }
 
 function useUnresolvedCashuSendQuotes() {
