+++ app/features/send/cashu-send-swap-hooks.ts
@@ -1,7 +1,4 @@
-import {
-  REALTIME_SUBSCRIBE_STATES,
-  type RealtimePostgresChangesPayload,
-} from '@supabase/supabase-js';
+import type { RealtimePostgresChangesPayload } from '@supabase/supabase-js';
 import {
   type QueryClient,
   useMutation,
@@ -12,6 +9,7 @@ import {
 } from '@tanstack/react-query';
 import { useEffect, useMemo, useState } from 'react';
 import type { Money } from '~/lib/money';
+import { useSupabaseRealtimeSubscription } from '~/lib/supabase/supabase-realtime';
 import { useLatest } from '~/lib/use-latest';
 import type { CashuAccount } from '../accounts/account';
 import {
@@ -175,15 +173,6 @@ function useSwapForProofsToSend() {
   });
 }
 
-type SubscriptionState =
-  | {
-      status: 'subscribing' | 'subscribed' | 'closed';
-    }
-  | {
-      status: 'error';
-      error: Error;
-    };
-
 function useOnCashuSendSwapChange({
   onCreated,
   onUpdated,
@@ -191,15 +180,12 @@ function useOnCashuSendSwapChange({
   onCreated: (swap: CashuSendSwap) => void;
   onUpdated: (swap: CashuSendSwap) => void;
 }) {
-  const [state, setState] = useState<SubscriptionState>({
-    status: 'subscribing',
-  });
   const encryption = useCashuCryptography();
   const onCreatedRef = useLatest(onCreated);
   const onUpdatedRef = useLatest(onUpdated);
 
-  useEffect(() => {
-    const channel = agicashDb
+  return useSupabaseRealtimeSubscription(() =>
+    agicashDb
       .channel('cashu-send-swaps')
       .on(
         'postgres_changes',
@@ -221,33 +207,8 @@ function useOnCashuSendSwapChange({
             onUpdatedRef.current(updatedSwap);
           }
         },
-      )
-      .subscribe((status, error) => {
-        if (status === REALTIME_SUBSCRIBE_STATES.SUBSCRIBED) {
-          setState({ status: 'subscribed' });
-        } else if (status === REALTIME_SUBSCRIBE_STATES.CLOSED) {
-          setState({ status: 'closed' });
-        } else {
-          setState({
-            status: 'error',
-            error: new Error(
-              `Error with "${channel.topic}" channel subscription. Status: ${status}`,
-              { cause: error },
-            ),
-          });
-        }
-      });
-
-    return () => {
-      channel.unsubscribe();
-    };
-  }, [encryption]);
-
-  if (state.status === 'error') {
-    throw state.error;
-  }
-
-  return state.status;
+      ),
+  );
 }
 
 export function useUnresolvedCashuSendSwaps() {
